Remove redis from tags

+++ commands/tags/delete.js
@@ -57,17 +57,9 @@ module.exports = class TagDeleteCommand extends Command {
 		const staffRole = this.client.isOwner(msg.author) || await msg.member.roles.exists('name', 'Server Staff');
 		const tag = await Tag.findOne({ where: { name, guildID: msg.guild.id } });
 		if (!tag) return msg.say(`A tag with the name **${name}** doesn't exist, ${msg.author}`);
-		if (tag.userID !== msg.author.id && !staffRole) {
-			return msg.say(`You can only delete your own tags, ${msg.author}`);
-		}
-		return Tag.sync()
-			.then(() => {
-				Tag.destroy({ where: { name, guildID: msg.guild.id } });
-				redis.db.delAsync(`tag${name}${msg.guild.id}`);
-				if (tag.example) {
-					msg.guild.channels.get(exampleChannel).fetchMessage(tag.exampleID).then(del => del.delete());
-				}
-				return msg.say(`The tag **${name}** has been deleted, ${msg.author}`);
-			});
+		if (tag.userID !== msg.author.id && !staffRole) return msg.say(`You can only delete your own tags, ${msg.author}`);
+		Tag.destroy({ where: { name, guildID: msg.guild.id } });
+		if (tag.example) msg.guild.channels.get(exampleChannel).fetchMessage(tag.exampleID).then(del => del.delete());
+		return msg.say(`The tag **${name}** has been deleted, ${msg.author}`);
 	}
 };
